Move i18nProvider to separate file
