+++ apps/admin/src/components/admin/Admin/Admin.tsx
@@ -1,6 +1,5 @@
 // import fakeDataProvider from 'ra-data-fakerest'
-import polyglotI18nProvider from 'ra-i18n-polyglot'
-import dutchMessages from 'ra-language-dutch'
+
 import { Admin as ReactAdmin, Resource } from 'react-admin'
 
 import { CategoryCreate } from '../../category/CategoryCreate/CategoryCreate'
@@ -12,6 +11,7 @@ import { FormList } from '../../form/FormList'
 
 import { CustomLayout } from './CustomLayout'
 import { dataProvider } from './dataProvider'
+import { i18nProvider } from './i18nProvider'
 // import { MainForm } from '../MainForm'
 
 // const dataProvider = fakeDataProvider({
@@ -71,37 +71,6 @@ import { dataProvider } from './dataProvider'
 //     },
 //   ],
 // })
-
-const nl = {
-  ...dutchMessages,
-  resources: {
-    classification: {
-      name: 'Categorie |||| Categorieën',
-      fields: {
-        form: 'Vragenlijst',
-        name: 'Naam',
-      },
-    },
-    form: {
-      name: 'Vragenlijst |||| Vragenlijsten',
-      fields: {
-        classification: 'Categorie',
-        title: 'Naam',
-      },
-    },
-  },
-  ma: {
-    dialog: {
-      overwriteClassification: {
-        content:
-          'Hier haal je de gekoppelde vragenlijst offline en zet je de huidige vragenlijst online. Is dat wat je wilt?',
-        title: 'Ontkoppel bestaande vragenlijst',
-      },
-    },
-  },
-}
-
-const i18nProvider = polyglotI18nProvider(() => nl, 'nl')
 
 export const Admin = () => (
   <ReactAdmin layout={CustomLayout} dataProvider={dataProvider()} i18nProvider={i18nProvider}>
